requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
black==22.3.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
kombu==5.2.4
lxml==4.8.0
Markdown==3.3.6
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.8.0
